+++ packages/cli-adapter/src/index.js
@@ -1,10 +1,12 @@
 import cliAdapter from './saihubot-cli-adapter';
 import {addonConfirm} from './saihubot-cli-addon-dialog.js';
 import {addonSearch, addonOpenLink} from './saihubot-cli-addon-search';
+import {skillHelp} from './saihubot-cli-addon-skill-help';
 
 export {
   cliAdapter,
   addonConfirm,
   addonSearch,
   addonOpenLink,
+  skillHelp,
 };